+++ src/autoroot-plugin/autoroot/autorootlinker.py
@@ -1,4 +1,4 @@
-from PyQt5.QtCore import QCoreApplication, qDebug, qInfo
+from PyQt5.QtCore import QCoreApplication, qDebug, qInfo, qWarning
 from pathlib import Path
 import mobase, os, json, pathlib, shutil
 
@@ -19,13 +19,13 @@ class AutoRootLinker():
             relativePath = self._paths.rootRelativePath(str(file))
             gamePath = self._paths.gamePath() / relativePath
             if gamePath.exists():
-                qInfo("AutoRoot: Renaming existing file " + str(gamePath))
+                qDebug("AutoRoot: Renaming existing file " + str(gamePath))
                 shutil.move(gamePath, Path(str(gamePath) + ".autoroot"))
-            qInfo("AutoRoot: Linking " + str(file) + " to " + str(gamePath))
+            qDebug("AutoRoot: Linking " + str(file) + " to " + str(gamePath))
             try:
                 Path(file).link_to(gamePath)
             except:
-                qInfo("AutoRoot: Could not create link " + str(gamePath))
+                qWarning("AutoRoot: Could not create link " + str(gamePath))
 
 
         qInfo("AutoRoot: Saving link data.")
@@ -37,20 +37,23 @@ class AutoRootLinker():
 
     def clear(self):
         qInfo("AutoRoot: Loading link data.")
-        linkFileData = json.load(open(self._paths.linkTempPath()))
-        qInfo("AutoRoot: Link data loaded.")
+        if self._paths.linkTempPath().exists():
+            linkFileData = json.load(open(self._paths.linkTempPath()))
+            qInfo("AutoRoot: Link data loaded.")
 
-        for file in linkFileData.keys():
-            gamePath = self._paths.gamePath() / file
-            if gamePath.exists():
-                qInfo("AutoRoot: Removing link " + str(gamePath))
-                gamePath.unlink(True)
-            if Path(str(gamePath) + ".autoroot").exists():
-                qInfo("AutoRoot: Restoring file " + str(gamePath))
-                shutil.move(Path(str(gamePath) + ".autoroot"), gamePath)
+            for file in linkFileData.keys():
+                gamePath = self._paths.gamePath() / file
+                if gamePath.exists():
+                    qDebug("AutoRoot: Removing link " + str(gamePath))
+                    gamePath.unlink(True)
+                if Path(str(gamePath) + ".autoroot").exists():
+                    qDebug("AutoRoot: Restoring file " + str(gamePath))
+                    shutil.move(Path(str(gamePath) + ".autoroot"), gamePath)
 
-        qInfo("AutoRoot: Clearing link data.")
-        os.remove(self._paths.linkTempPath())
-        qInfo("AutoRoot: Link data cleared.")
+            qInfo("AutoRoot: Clearing link data.")
+            os.remove(self._paths.linkTempPath())
+            qInfo("AutoRoot: Link data cleared.")
+        else:
+            qInfo("AutoRoot: No link file exists.")
         
     #endregion
